Separated factories into separate source files
