+++ newsfragments/+c76e72c5.misc.rst
@@ -1,0 +1,1 @@
+Separated factories into its own source files.
